Add RoomTriggers that close doors
When the room is entered by the whole team, the HostEncounterManager calls BeginEncounter and the doors close.

+++ Assets/Scripts/Common/LevelGeneration/RoomFacade.cs
@@ -7,7 +7,7 @@ using Zenject;
 /// </summary>
 public class RoomFacade: MonoUpdatableObject
 {
-    public int ID { get; set; }
+    public ushort ID { get; set; }
 
     public override void OnUpdate(float deltaTime)
     {
